(refactor) show either mentee or mentor while rendering a single profile
