+++ client/src/App.js
@@ -3,8 +3,8 @@ import React from 'react'
 import './App.css'
 import Register from './Pages/Register'
 import Login from './Pages/Login'
-import MenteeProfile from "./Pages/MenteeProfile"
-import MentorProfile from "./Pages/MentorProfile"
+import Profile from "./Pages/MenteeProfile"
+import MentorProfile from "./Pages/Profile"
 import Landing from './Pages/Landing'
 import Homepage from './Pages/Homepage'
 import Settings from './Pages/Settings'
@@ -68,7 +68,7 @@ function App () {
             </Route>
             <Route path='/login'>
               {!state.isAuthenticated ? <Login /> : 
-                state.userType === "mentor"? <MentorProfile/>: <MenteeProfile/>
+               <MentorProfile/>
               }
               {console.log(state.userType)}
             </Route>
@@ -76,10 +76,10 @@ function App () {
               <Homepage />
             </Route>
             <Route path='/profile/mentee'>
-              {!state.isAuthenticated && state.userType === "mentee"? <Login /> : <MenteeProfile />}
+              {!state.isAuthenticated && state.userType === "mentee"? <Login /> : <Profile />}
             </Route>
             <Route path='/profile/mentor'>
-              {!state.isAuthenticated && state.userType === "mentor" ? <Login /> : <MentorProfile />}
+              {!state.isAuthenticated && state.userType === "mentor" ? <Login /> : <Profile />}
             </Route>
             <Route path='/settings'>
               <Settings isMentor={false}/>
